==> notebook_file/1.KBL_crawling.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import Select

import time
import pandas as pd

driver = webdriver.Chrome('chromedriver')
driver.get('https://www.kbl.or.kr/game/archive-player')
driver.maximize_window()
driver.implicitly_wait(3)

#모든 정보 저장할 df
df = pd.DataFrame()

#경기 종류 선택
select_type = Select(driver.find_element_by_xpath('//*[@id="container"]/main/div[2]/div[1]/div/div[1]/div/select'))
options_type = [o.text for o in driver.find_elements_by_xpath('//*[@id="container"]/main/div[2]/div[1]/div/div[1]/div/select/option')]

for o_type in options_type:
    select_type.select_by_visible_text(o_type)
    time.sleep(2)
    
    #년도 선택
    select_year = Select(driver.find_element_by_xpath('//*[@id="container"]/main/div[2]/div[1]/div/div[2]/div/select'))
    options_year = [o.text for o in driver.find_elements_by_xpath('//*[@id="container"]/main/div[2]/div[1]/div/div[2]/div/select/option')]
    
    for year in options_year:
        select_year.select_by_visible_text(year)
        time.sleep(2)
        
        #선수정보 df 생성
        columns_header = ['경기', '년도', '순위', '선수명', '팀']
        header = driver.find_element_by_xpath('//*[@id="container"]/main/div[2]/div[2]/section/section[1]/div/div/table')
        no = [no.text for no in header.find_elements_by_class_name('no')]
        name = [name.text for name in header.find_elements_by_class_name('name')]
        team = [team.text for team in header.find_elements_by_class_name('team')]
        player = [[o_type, year, i, j, k] for i, j, k in zip(no[1:], name[1:], team[1:])]
        df_header = pd.DataFrame(player, columns=columns_header)

        #data df 생성
        columns_data = ['PTS', 'FG', 'FGA', 'FG%', 'FT', 'FTA', 'FT%', 'PP', 'PPA', 'PP%', 'REB_OFF', 'REB_DEF', 'REB_TOT', 'AST', 'TO', 'STL', 'BS', 'DF']
        data = driver.find_element_by_xpath('//*[@id="container"]/main/div[2]/div[2]/section/section[1]/div/div/div/table/tbody')
        data = data.text.split('\n')
        data = [d.split(' ') for d in data]
        df_data = pd.DataFrame(data, columns=columns_data)
        
        #선수정보 df, data df concat
        df_total = pd.concat([df_header, df_data], axis=1)
        
        #통합 df에 concat
        df = pd.concat([df, df_total])
        print("경기종류: {}, 년도: {} succesfully crawled ----- # of data: {}: ".format(o_type, year, df.shape[0]))

#index 초기화
df = df.reset_index(drop=True)
        
#결측치 확인
df.isnull().values.any()

#csv로 저장
df.to_csv('KBL.csv', encoding='utf-8-sig')
