==> notebook_file/2.KBL_preprocessing.ipynb <==
MAX_YEAR = 4

import pandas as pd

#크롤링한 데이터 읽어오기
df = pd.read_csv('KBL.csv')
df = df.sort_values(by=['선수명', '년도'], axis=0, ascending=True)
df = df.reset_index(drop=True)

#데뷔년도부터 최대 MAX_YEAR만큼의 데이터만 유지
unique = set()
for idx, row in df.iterrows():
    if row['선수명'] not in unique:
        unique.add(row['선수명'])
        first_year = row['년도']
        first_year = int(first_year[:4])
    max_year = first_year + MAX_YEAR - 1
    max_year = str(max_year) + "-" + str(max_year + 1)
    if row['년도'] > max_year: df = df.drop(index=idx)
        
#index 초기화
df = df.reset_index(drop=True)

#csv로 저장
df.to_csv('KBL_preprocessed.csv', encoding='utf-8-sig')
