==> notebook_file/5.preprocess.ipynb <==
import pandas as pd
import numpy as np

df1 = pd.read_csv('KBL_preprocessed.csv')
df2 = pd.read_csv('프로기록있는애들.csv')
df3 = pd.read_csv('프로기록없는애들.csv')

#normalization 대상 columns
df1normcol = df1.columns[7:]
df2normcol = df2.columns[2:34]
df3normcol = df3.columns[2:34]

fd1 = df1[df1normcol].to_numpy()
fd2 = df2[df2normcol].to_numpy()
fd3 = df3[df3normcol].to_numpy()

def normalize(df):
    mean = df.mean(axis=0)
    mean = mean.reshape(1,mean.shape[0])
    std = df.std(axis=0)
    std= std.reshape(1,std.shape[0])
    newdf = df - mean
    newdf = newdf/std
    return newdf

def nancoldel(df):
    threshold = df.shape[0] * 0.7
    for col in df.columns :
        counter = df[col].count()
        if counter<threshold:
            df = df.drop(col,axis=1)
    newdf = df.dropna(axis=0)
    newdf = newdf.reset_index(drop=True)
    return newdf

#날짜, percentage 데이터 전처리
for i in range(fd2.shape[0]):
    k= str(fd2[i,1])
    k = k.replace('\'','')
    k = k.replace(':','')
    val = int(k[:2])*3600 + int(k[2:4])*60 +int(k[4:])
    fd2[i,1] = val
    l = str(fd2[i,6]).replace('\'','')
    m = str(fd2[i,9]).replace('\'','')
    n = str(fd2[i,12]).replace('\'','')
    l = l.replace('.','')
    m = m.replace('.','')
    n = n.replace('.','')
    l = int(l.replace('%',''))
    m = int(m.replace('%',''))
    n = int(n.replace('%',''))
    
    fd2[i,6] = l
    fd2[i,9] = m
    fd2[i,12] = n
    
for i in range(fd3.shape[0]):
    k= str(fd3[i,1])
    k = k.replace('\'','')
    k = k.replace(':','')
    val = int(k[:2])*3600 + int(k[2:4])*60 +int(k[4:])
    fd3[i,1] = val
    l = str(fd3[i,6]).replace('\'','')
    m = str(fd3[i,9]).replace('\'','')
    n = str(fd3[i,12]).replace('\'','')
    l = l.replace('.','')
    m = m.replace('.','')
    n = n.replace('.','')
    l = int(l.replace('%',''))
    m = int(m.replace('%',''))
    n = int(n.replace('%',''))
    fd3[i,6] = l
    fd3[i,9] = m
    fd3[i,12] = n

#std에서 에러나길래 타입 통일해줌
fd1 = fd1.astype(np.float32)
fd2 = fd2.astype(np.float32)
fd3 = fd3.astype(np.float32)

fd1 = normalize(fd1)
fd2 = normalize(fd2)
fd3 = normalize(fd3)

normeddf1 = pd.DataFrame(fd1)
normeddf2 = pd.DataFrame(fd2)
normeddf3 = pd.DataFrame(fd3)

df1col = df1.columns[2:7] 
df2col1 = df2.columns[:2]
df3col1 = df3.columns[:2]
df2col2 = df2.columns[34:]
df3col2 = df3.columns[34:]

newdf1 = pd.concat([df1[df1col],normeddf1],axis=1)
newdf2 = pd.concat([df2[df2col1],normeddf2],axis=1)
newdf2 = pd.concat([newdf2[newdf2.columns[1:]], df2[df2col2]],axis=1)
newdf3 = pd.concat([df3[df3col1],normeddf3],axis=1)
newdf3 = pd.concat([newdf3[newdf3.columns[1:]], df3[df3col2]],axis=1)

newdf1.columns = ['경기', '년도', '순위', '선수명', '팀', 'PTS', 'FG', 'FGA', 'FG%', 'FT', 'FTA',
       'FT%', 'PP', 'PPA', 'PP%', 'REB_OFF', 'REB_DEF', 'REB_TOT', 'AST', 'TO',
       'STL', 'BS', 'DF']
newdf2.columns = ['선수명', '경기수', '전체 출전시간', '득점 전체', '득점 평균', '2점슛 성공', '2점슛 시도',
       '2점슛 성공률', '3점슛 성공', '3점슛 시도', '3점슛 성공률', '자유투 성공', '자유투 시도', '자유투 성공률',
       '공격 리바운드', '수비 리바운드', '평균 리바운드', '전체 어시스트', '평균 어시스트', '스틸', '블록',
       '덩크슛 성공', '덩크슛 시도', '파울(자유튜유)', '파울(자유투무)', 'IDF', '테크니컬 파울', '굿디펜스',
       '속공', 'FBF', '턴오버', '팀 리바운드', '팀 턴오버', '대회년도', '대학교']
newdf3.columns = ['선수명', '경기수', '전체 출전시간', '득점 전체', '득점 평균', '2점슛 성공', '2점슛 시도',
       '2점슛 성공률', '3점슛 성공', '3점슛 시도', '3점슛 성공률', '자유투 성공', '자유투 시도', '자유투 성공률',
       '공격 리바운드', '수비 리바운드', '평균 리바운드', '전체 어시스트', '평균 어시스트', '스틸', '블록',
       '덩크슛 성공', '덩크슛 시도', '파울(자유튜유)', '파울(자유투무)', 'IDF', '테크니컬 파울', '굿디펜스',
       '속공', 'FBF', '턴오버', '팀 리바운드', '팀 턴오버', '대회년도', '대학교']

newdf1 = nancoldel(newdf1)
newdf2 = nancoldel(newdf2)
newdf3 = nancoldel(newdf3)

newdf1.to_csv('kbl_normalized.csv', encoding='utf-8-sig')
newdf2.to_csv('univtopro_normalized.csv', encoding='utf-8-sig')
newdf3.to_csv('univnopro_normalized.csv', encoding='utf-8-sig')
