==> notebook_file/8.try_Regression.ipynb <==
import pandas as pd
import numpy as np

from sklearn.linear_model import LinearRegression, LogisticRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import r2_score

univ = pd.read_csv('univtopro_preprocessed.csv')
pro = pd.read_csv('KBL.csv')

univ_avg = univ.groupby(univ['선수명']).mean()
pro_avg = pro.groupby(pro['선수명']).mean()

data = pd.merge(univ_avg, pro_avg, how='inner', left_index=True, right_index=True)
data

#가산점 항목
KBL_eff_pos = (data['PTS'].to_numpy() + data['STL'].to_numpy() + data['BS'].to_numpy() + data['REB_DEF'].to_numpy()) * 1.0\
            + (data['REB_OFF'].to_numpy() + data['AST'].to_numpy()) * 1.5

#감점 항목
KBL_eff_neg = (data['TO'].to_numpy()) * 1.5\
            + (data['FGA'].to_numpy() - data['FG'].to_numpy()) * 1.0\
            + (data['PP'].to_numpy() - data['PP'].to_numpy()) * 1.0\
            + (data['FTA'].to_numpy() - data['FT'].to_numpy()) * 0.8\

KBL_eff = KBL_eff_pos - KBL_eff_neg

#-1과 1사이 수치로 정규화
#KBL_eff_max, KBL_eff_min = np.max(KBL_eff), np.min(KBL_eff)
#KBL_eff = KBL_eff / KBL_eff_max if abs(KBL_eff_max) > abs(KBL_eff_min) else KBL_eff / -KBL_eff_min

data['KBL_EFF'] = KBL_eff
data

X = data[['경기수', '전체 출전시간', '득점 전체', '득점 평균', '2점슛 성공', '2점슛 시도',
       '2점슛 성공률', '3점슛 성공', '3점슛 시도', '3점슛 성공률', '자유투 성공', '자유투 시도', '자유투 성공률',
       '공격 리바운드', '수비 리바운드', '평균 리바운드', '전체 어시스트', '평균 어시스트', '스틸', '블록',
       '덩크슛 성공', '덩크슛 시도', '파울(자유튜유)', '파울(자유투무)', '테크니컬 파울', '굿디펜스',
       '속공', 'FBF', '턴오버', '팀 리바운드', '팀 턴오버']]

y = data['KBL_EFF']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=0)

linear_regressor = LinearRegression()
linear_regressor.fit(X_train, y_train)
y_pred = linear_regressor.predict(X_test)
r2_score(y_pred, y_test)
